reorganize base tests
